+++ app/src/main/java/com/example/noteapp/domain/repository/NoteRepository.kt
@@ -8,6 +8,6 @@ interface NoteRepository {
     suspend fun addNote(noteEntity: NoteEntity)
     suspend fun updateNote(noteEntity: NoteEntity)
     suspend fun deleteNote(noteEntity: NoteEntity)
-    fun getNoteById(id: Int): Flow<NoteEntity>
+    fun getNoteById(id: Long): Flow<NoteEntity>
     suspend fun getLastNoteId(): Long?
 }
